updatevpnbookpass.1d.sh
Hello!
My first time submitting plugins to BitBar so hope everything is in order for you.
I've created a plugin which I'd like to submit:
"Automatically update vpn password." The password is collected from the VPNBOOK page.
It requires some dependencies:  
set up PPTP VPN from the following page: https://www.vpnbook.com
and show PPTP status in the menu bar

We have to manually give the rights to this script:
*  Launch terminal and go inside the bitbar plugins folder
*  chmod 751 updatevpnbookpass.1d.sh

Some commands of the script needs admin rights. It can be done adding the name of the script to the parameter file 'sudoers', giving the possibility to anybody to execute it:
*  Launch terminal
*  sudo su - root (to connect with root user)
*  (enter the root password)
*  sudo visudo
*  <your user>      ALL= NOPASSWD: <your path>/updatevpnbookpass.1d.sh
(This line is to add in the User privilege specification block)
(where <your user> is your session account
(and <your path> is the directory where you place your bitbar plugins)
*  :wq! (to leave visudo)

Please let me know if any changes are required before you can merge this pull request. Thanks!

Joce